ProThreadedAlertService for alert services that support resolving alerts
Ticket: #26679
